Implement filter_plugin to manage ESI transformation
