+++ ansible_collections/arista/avd/roles/eos_l3ls_evpn/README.md
@@ -1245,7 +1245,7 @@ servers:
           short_esi: 0303:0202:0101
 ```
 
-`short_esi` is an abreviated 3 octets value to encode [Ethernet Segment ID](https://tools.ietf.org/html/rfc7432#section-8.3.1) and LACP ID. Below is the result of `short_esi` transformation to network values:
+`short_esi` is an abreviated 3 octets value to encode [Ethernet Segment ID](https://tools.ietf.org/html/rfc7432#section-8.3.1) and LACP ID. Transformation from abstraction to network values is managed by a [filter_plugin](../plugins/filter/esi_management.py) and provides following result:
 
 - _EVPN ESI_: 000:000:0303:0202:0101
 - _LACP ID_: 0303.0202.0101
