+++ ansible_collections/arista/avd/roles/eos_l3ls_evpn/templates/edge-ports/leaf-server-port-channels.j2
@@ -18,10 +18,10 @@
 {# Add ESI Support isse aristanetworks/ansible-avd #215 #}
 {%                     if adapter.switches | unique | list | length > 1 %}
 {%                         if adapter.port_channel.short_esi is defined and adapter.port_channel.short_esi is not none and adapter.port_channel.short_esi.split(':') | length == 3 %}
-    esi: 0000:0000:{{ adapter.port_channel.short_esi }}
-    rt: {{ adapter.port_channel.short_esi | arista.avd.esi_to_rt }}
+    esi: {{ adapter.port_channel.short_esi | arista.avd.generate_esi }}
+    rt: {{ adapter.port_channel.short_esi | arista.avd.generate_route_target }}
 {%                             if adapter.port_channel.mode is defined and adapter.port_channel.mode == 'active' %}
-    lacp_id: {{ adapter.port_channel.short_esi.replace(':', '.') }}
+    lacp_id: {{ adapter.port_channel.short_esi | arista.avd.generate_lacp_id }}
 {%                             endif %}
 {%                         endif %}
 {%                     endif %}
